requirements.txt
Django==1.8
django-cleanup==2.1.0
django-contrib-comments==1.8.0
django-crispy-forms==1.7.2
django-threadedcomments==1.1
djangorestframework==3.1.0
djangotoolbox==1.8.0
Pillow==5.0.0
